provide better structur via grouping classes by modules
