+++ lib/reports/console.rb
@@ -2,42 +2,44 @@
 
 require_relative '../calculations/drawdown'
 
-# simple console reporter
-class Console
-  attr_reader :data
+module Reports
+  # simple console reporter
+  class Console
+    attr_reader :data
 
-  def initialize(data)
-    @data = data
-  end
-
-  def generate_report
-    return unless data && data['data']
-    data['data'].each do |day|
-      color_puts("#{day[0]}: Closed at #{day[4]} (#{day[3]} ~ #{day[2]})", 32)
+    def initialize(data)
+      @data = data
     end
-    first_drawdowns
-    maximum_drawdown
-    return_summary
-  end
 
-  def first_drawdowns
-    Drawdown.new(data).drawdowns.each do |drawdown|
-      color_puts(
-        "#{drawdown[1]} on #{drawdown[0]} -> #{drawdown[2]} on #{drawdown[3]}",
-        31
-      )
+    def generate_report
+      return unless data && data['data']
+      data['data'].each do |day|
+        color_puts("#{day[0]}: Closed at #{day[4]} (#{day[3]} ~ #{day[2]})", 32)
+      end
+      first_drawdowns
+      maximum_drawdown
+      return_summary
     end
-  end
 
-  def maximum_drawdown
-  end
+    def first_drawdowns
+      Calculations::Drawdown.new(data).drawdowns.each do |drawdown|
+        color_puts(
+          "#{drawdown[1]} on #{drawdown[0]} -> #{drawdown[2]} on #{drawdown[3]}",
+          31
+        )
+      end
+    end
 
-  def return_summary
-  end
+    def maximum_drawdown
+    end
 
-  private
+    def return_summary
+    end
 
-  def color_puts(text, color)
-    puts "\033[#{color}m #{text}\033[0m"
+    private
+
+    def color_puts(text, color)
+      puts "\033[#{color}m #{text}\033[0m"
+    end
   end
 end
